FUSETOOLS2-82 - extracting a CamelKafkaConnectUtility class
