+++ src/main/java/com/github/cameltooling/lsp/internal/parser/CamelKafkaConnectDSLParser.java
@@ -34,33 +34,26 @@ import com.github.cameltooling.lsp.internal.instancemodel.PropertiesDSLModelHelp
  */
 public class CamelKafkaConnectDSLParser extends ParserFileHelper {
 	
-	public static final String CAMEL_SINK_URL = "camel.sink.url";
-	public static final String CAMEL_SOURCE_URL = "camel.source.url";
 	private static final Logger LOGGER = LoggerFactory.getLogger(CamelKafkaConnectDSLParser.class);
 
 	@Override
 	public String getCamelComponentUri(String line, int characterPosition) {
-		if (isInsideACamelUri(line, characterPosition)) {
+		if (new CamelKafkaConnectUtility().isInsideACamelUri(line, characterPosition)) {
 			Properties properties = new Properties();
 			try {
 				properties.load(new StringReader(line));
 
-				String camelComponentURI = properties.getProperty(CAMEL_SOURCE_URL);
+				String camelComponentURI = properties.getProperty(CamelKafkaConnectUtility.CAMEL_SOURCE_URL);
 				if (camelComponentURI != null) {
 					return camelComponentURI;
 				} else {
-					return properties.getProperty(CAMEL_SINK_URL);
+					return properties.getProperty(CamelKafkaConnectUtility.CAMEL_SINK_URL);
 				}
 			} catch (IOException e) {
 				LOGGER.warn("Cannot retrieve Camel Component URI",e);
 			}
 		}
 		return null;
-	}
-
-	public boolean isInsideACamelUri(String line, int characterPosition) {
-		return line.startsWith(CAMEL_SOURCE_URL) && CAMEL_SOURCE_URL.length() < characterPosition 
-				|| line.startsWith(CAMEL_SINK_URL) && CAMEL_SINK_URL.length() < characterPosition;
 	}
 
 	@Override
@@ -86,9 +79,9 @@ public class CamelKafkaConnectDSLParser extends ParserFileHelper {
 
 	public String getCorrespondingMethodName(TextDocumentItem textDocumentItem, int lineNumber) {
 		String line = parserFileHelperUtil.getLine(textDocumentItem, lineNumber);
-		if (line.startsWith(CAMEL_SINK_URL)) {
+		if (line.startsWith(CamelKafkaConnectUtility.CAMEL_SINK_URL)) {
 			return "to";
-		} else if(line.startsWith(CAMEL_SOURCE_URL)) {
+		} else if(line.startsWith(CamelKafkaConnectUtility.CAMEL_SOURCE_URL)) {
 			return "from";
 		}
 		return null;
